requirements.txt
python-telegram-bot
aiohttp
tzdata
splitwise
aiopvpc
seaborn
translate
tinytuya
